refactor: separate notification and device logic into resource-based architecture
- Create NotificationResource class with full CRUD operations
  - Paginated queries with filters (status, type, category, priority)
  - Mark as read functionality (single, multiple, all)
  - Statistics and preferences management
  - Helper methods (getUnread, getByType, getUrgent, etc.)

- Create DeviceResource class for device management
  - Device registration and deregistration
  - FCM token updates
  - Device listing with filters (active, inactive, by type)
  - Helper methods (getMobileDevices, isDeviceRegistered, getCount)

- Create FirebaseMessagingService for Firebase-specific operations
  - FCM token management and initialization
  - Device UUID generation and storage
  - Browser/platform detection
  - Real-time message listeners
  - Foreground notification display
  - Uses DeviceResource for backend API calls

- Update components to use new architecture
  - NotificationBell: Uses NotificationResource, DeviceResource, firebaseMessaging
  - Notifications page: Uses NotificationResource, DeviceResource, firebaseMessaging

- Remove monolithic notification.service.ts
- Fix TypeScript duplicate type exports (PaginatedResponse)

Benefits:
- Separation of concerns (Firebase vs API logic)
- Consistency with existing resource patterns
- Better type safety and code organization
- Improved maintainability and reusability
