+++ frontend/src/lib/resources/index.ts
@@ -2,6 +2,8 @@
 export * from './user';
 export * from './department';
 export * from './jobPosition';
+export * from './notification';
+export * from './device';
 
 // Re-export the main API client
 export { apiClient, TokenManager } from '../api';
